requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.23.2
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv==1.0.1
SQLAlchemy==1.4.51
alembic==1.11.1
psycopg2-binary~=2.9.3